refactor build system

+++ engine/core/engine.cpp
@@ -126,7 +126,7 @@ namespace engine::core
     }
     Engine::Engine() 
     {
-        random_engine_ = std::make_unique<std::mt19937>(random_seed());
+        random_engine_ = std::make_unique<std::mt19937>(static_cast<uint32_t>(random_seed()));
         event_function_ = std::bind_front([this](Event &e) __lambda_force_inline
                                           { OnEvent(e); });
         std::filesystem::path target_folder;
